+++ AzLandingZone/Private/rest-api/sentinel/Set-LzFusionAlertRule.ps1
@@ -1,0 +1,41 @@
+Function Set-LzFusionAlertRule {
+	Param (
+		[String] $name = "lzslz"
+	)
+
+	if(!($GetResourceGroup = Get-AzResourceGroup -ResourceGroupName "*$name*")){
+		Write-Host "No Resource Group for Secure Landing Zone found"
+		Write-Host "Please run setup script before running this script"
+		return 1;
+	}
+	if(!($GetLogAnalyticsWorkspace = Get-AzOperationalInsightsWorkspace -ResourceGroupName $GetResourceGroup.ResourceGroupName)){
+		Write-Host "No Log analytics workspace for Secure Landing Zone found"
+		Write-Host "Please run setup script before running this script"
+		return 1;
+	}
+
+	$body = @{
+		'etag'       = ''
+		'kind'       = 'Fusion'
+		'properties' = @{
+			'enabled' = 'true'
+			'alertRuleTemplateName' = 'f71aba3d-28fb-450b-b192-4e76a83015c8'
+		}
+	}
+	$uri = "https://management.azure.com" + $GetLogAnalyticsWorkspace.ResourceId  + "/providers/Microsoft.SecurityInsights/alertRules/f71aba3d-28fb-450b-b192-4e76a83015c8?api-version=2020-01-01"
+
+	try {
+		Write-Host -ForegroundColor Green "Enabling alerts creation for Azure Advanced Threat Protection in Sentinel"
+		$auth = Get-LzAccessToken
+		$requestResult = Invoke-webrequest -Uri $uri -Method Put -Headers $auth -Body ($body | ConvertTo-Json -Depth 5)
+		Write-Host -ForegroundColor Green "Enabled alerts creation for Azure Advanced Threat Protection in Sentinel"
+	}
+	catch {
+		Switch ($_.Exception.Response.StatusCode.value__)
+		{
+			409 {Write-Host "Alerts creation for Azure Advanced Threat Protection in Sentinel already enabled" -ForegroundColor Yellow}
+			default {Write-Host "An unexpected error happened. Contact Landing Zone FMB for additional support." -ForegroundColor Red}
+		}
+	}
+}
+Export-ModuleMember -Function Set-LzAzureATPAlertRule