+++ AzLandingZone/Private/rest-api/sentinel/Set-LzMicrosoftCloudAppSecurityAlertRule.ps1
@@ -1,0 +1,44 @@
+Function Set-LzMicrosoftCloudAppSecurityAlertRule {
+	Param (
+		[String] $name = "lzslz"
+	)
+
+	if(!($GetResourceGroup = Get-AzResourceGroup -ResourceGroupName "*$name*")){
+		Write-Host "No Resource Group for Secure Landing Zone found"
+		Write-Host "Please run setup script before running this script"
+		return 1;
+	}
+	if(!($GetLogAnalyticsWorkspace = Get-AzOperationalInsightsWorkspace -ResourceGroupName $GetResourceGroup.ResourceGroupName)){
+		Write-Host "No Log analytics workspace for Secure Landing Zone found"
+		Write-Host "Please run setup script before running this script"
+		return 1;
+	}
+
+	$body = @{
+		'etag'       = ''
+		'kind'       = 'MicrosoftSecurityIncidentCreation'
+		'properties' = @{
+			'productFilter' = 'Microsoft Cloud App Security'
+			'displayName' = 'Create incidents based on Microsoft Cloud App Security alerts'
+            'enabled' = 'true'
+            'description' = 'Create incidents based on all alerts generated in Microsoft Cloud App Security'
+            'alertRuleTemplateName' = 'b3cfc7c0-092c-481c-a55b-34a3979758cb'
+		}
+	}
+	$uri = "https://management.azure.com" + $GetLogAnalyticsWorkspace.ResourceId  + "/providers/Microsoft.SecurityInsights/alertRules/b3cfc7c0-092c-481c-a55b-34a3979758cb?api-version=2020-01-01"
+
+	try {
+		Write-Host -ForegroundColor Green "Enabling alerts creation for Microsoft Cloud App Security in Sentinel"
+		$auth = Get-LzAccessToken
+		$requestResult = Invoke-webrequest -Uri $uri -Method Put -Headers $auth -Body ($body | ConvertTo-Json -Depth 5)
+		Write-Host -ForegroundColor Green "Enabled alerts creation for Microsoft Cloud App Security in Sentinel"
+	}
+	catch {
+		Switch ($_.Exception.Response.StatusCode.value__)
+		{
+			409 {Write-Host "Alerts creation for Microsoft Cloud App Security in Sentinel already enabled" -ForegroundColor Yellow}
+			default {Write-Host "An unexpected error happened. Contact Landing Zone FMB for additional support." -ForegroundColor Red}
+		}
+	}
+}
+Export-ModuleMember -Function Set-LzMicrosoftCloudAppSecurityAlertRule